Time: 0 ms (100%), Space: 11.6 MB (82.16%) - LeetHub

+++ 0003-longest-substring-without-repeating-characters/0003-longest-substring-without-repeating-characters.cpp
@@ -3,8 +3,8 @@ public:
     int lengthOfLongestSubstring(string s) {
         vector<int> seen(256,-1);
         int start = 0;
-        int best = 0;
-        for (int i=0; i<s.size(); i++){
+        int best  = 0;
+        for (int i=0; i<(int)s.size(); i++){
             char ch = s[i];
             if (seen[ch] != -1)
                 start = max(start,seen[ch]+1);
